Fix invalid use of getTreeNodeSchemaFromHydratedFlexNode
