--- packages/dds/tree/src/simple-tree/core/schemaCaching.ts
@@ -1,35 +1,0 @@
-/*!
- * Copyright (c) Microsoft Corporation and contributors. All rights reserved.
- * Licensed under the MIT License.
- */
-
-import { assert } from "@fluidframework/core-utils/internal";
-
-import { fail } from "../../util/index.js";
-
-import type { TreeNodeSchema } from "./treeNodeSchema.js";
-import type { InnerNode } from "./treeNodeKernel.js";
-import { UnhydratedFlexTreeNode } from "./unhydratedFlexTree.js";
-import { SimpleContextSlot, type Context } from "./context.js";
-
-/**
- * Gets the {@link TreeNodeSchema} for the {@link InnerNode}.
- */
-export function getSimpleNodeSchemaFromInnerNode(innerNode: InnerNode): TreeNodeSchema {
-	const context: Context = getSimpleContextFromInnerNode(innerNode);
-	return context.schema.get(innerNode.schema) ?? fail(0xb3f /* missing schema from context */);
-}
-
-/**
- * Gets the {@link Context} for the {@link InnerNode}.
- */
-export function getSimpleContextFromInnerNode(innerNode: InnerNode): Context {
-	if (innerNode instanceof UnhydratedFlexTreeNode) {
-		return innerNode.simpleContext;
-	}
-
-	const context = innerNode.anchorNode.anchorSet.slots.get(SimpleContextSlot);
-	assert(context !== undefined, 0xa55 /* missing simple tree context */);
-
-	return context;
-}
