+++ packages/dds/tree/src/simple-tree/core/treeNodeKernel.ts
@@ -423,22 +423,25 @@ export function tryDisposeTreeNode(anchorNode: AnchorNode): void {
 }
 
 /**
- * Lookup a TreeNodeSchema from a Hydrated FlexTreeNode.
- * @privateRemarks
- * This provides a way to access simple tree schema from the flex tree without depending on {@link FlexTreeSchema} which is in the process of being removed.
- * This is currently limited to hydrated nodes: this limitation will have to be fixed before {@link FlexTreeSchema} can be fully removed.
+ * Gets the {@link TreeNodeSchema} for the {@link InnerNode}.
  */
-export function getTreeNodeSchemaFromHydratedFlexNode(flexNode: FlexTreeNode): TreeNodeSchema {
-	assert(
-		flexNode.context.isHydrated(),
-		0xa56 /* getTreeNodeSchemaFromHydratedFlexNode only allows hydrated flex tree nodes */,
-	);
+export function getSimpleNodeSchemaFromInnerNode(innerNode: InnerNode): TreeNodeSchema {
+	const context: Context = getSimpleContextFromInnerNode(innerNode);
+	return context.schema.get(innerNode.schema) ?? fail(0xb3f /* missing schema from context */);
+}
 
-	const context =
-		flexNode.anchorNode.anchorSet.slots.get(SimpleContextSlot) ??
-		fail(0xb43 /* Missing SimpleContextSlot */);
+/**
+ * Gets the {@link Context} for the {@link InnerNode}.
+ */
+export function getSimpleContextFromInnerNode(innerNode: InnerNode): Context {
+	if (innerNode instanceof UnhydratedFlexTreeNode) {
+		return innerNode.simpleContext;
+	}
 
-	return context.schema.get(flexNode.schema) ?? fail(0xb44 /* Missing schema */);
+	const context = innerNode.anchorNode.anchorSet.slots.get(SimpleContextSlot);
+	assert(context !== undefined, 0xa55 /* missing simple tree context */);
+
+	return context;
 }
 
 /**
@@ -491,8 +494,11 @@ export function treeNodeFromAnchor(anchorNode: AnchorNode): TreeNode | TreeValue
  * This does not do caching or validation: caller must ensure duplicate nodes for a given inner node are not created, and that the inner node is valid.
  */
 export function createTreeNodeFromInner(innerNode: InnerNode): TreeNode | TreeValue {
-	const classSchema = getTreeNodeSchemaFromHydratedFlexNode(innerNode);
+	const classSchema = getSimpleNodeSchemaFromInnerNode(innerNode);
+	const internal = innerNode as unknown as InternalTreeNode;
 	return typeof classSchema === "function"
-		? new classSchema(innerNode as unknown as InternalTreeNode)
-		: (classSchema as { create(data: InnerNode): TreeValue }).create(innerNode);
+		? new classSchema(internal)
+		: (classSchema as { create(data: InternalTreeNode): TreeNode | TreeValue }).create(
+				internal,
+			);
 }
